requirements.txt
flask
prettytable
pytz
watchdog
dotenv
requests
pymysql
sqlalchemy
apscheduler
flask_socketio